removing warnings
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@4295 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/trajectory/BlendedTrajectory.cpp
@@ -73,7 +73,7 @@ BlendedTrajectory<T>::BlendedTrajectory(rw::models::Device::Ptr deviceIn,
     device = deviceIn;
     // Get the other parameters
     path = pathIn;
-    Npath = path.size();
+    Npath = (unsigned int) path.size();
     betas = betaIn;
     vscale = vscaleIn;
     ascale = ascaleIn;
@@ -269,7 +269,7 @@ template<class T>
 bool BlendedTrajectory<T>::init() {
     if(verbose) std::cout << "Initializing..." << std::endl;
     // Get DOF
-    K = device->getDOF();
+    K = (unsigned int) device->getDOF();
     // Initialize all vectors
     aSMax.m().resize(K);
     aSMin.m().resize(K);
